requirements.txt
aiohttp==3.10.11
dataframe_image==0.2.7
datasets==3.1.0
fire==0.7.1
huggingface_hub==0.36.0
ipython==8.12.3
numpy>=1.25,<2.0
pandas==2.0.3
Pillow==12.0.0
prompt_toolkit==3.0.52
python-dotenv==1.2.1
pytz==2025.2
setuptools==68.2.0
tqdm==4.67.1
typer==0.20.0
openai==2.9.0
loguru==0.7.3
path==17.1.1
